Implement DictionaryGenerator
DictionaryGenerator replaces HashSetDemo using the more efficient BufferedReader.
